+++ src/test/java/be/ugent/rml/TestCore.java
@@ -25,7 +25,11 @@ public abstract class TestCore {
     final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     Executor createExecutor(String mapPath) throws Exception {
-        return createExecutor(mapPath, new ArrayList<>());
+        return createExecutor(mapPath, new ArrayList<>(), null);
+    }
+
+    Executor createExecutor(String mapPath, List<Quad> extraQuads) throws Exception {
+        return createExecutor(mapPath, extraQuads, null);
     }
 
     /**
@@ -36,7 +40,7 @@ public abstract class TestCore {
      * @return An executor.
      * @throws Exception
      */
-    Executor createExecutor(String mapPath, List<Quad> extraQuads) throws Exception {
+    Executor createExecutor(String mapPath, List<Quad> extraQuads, String parentPath) throws Exception {
         ClassLoader classLoader = getClass().getClassLoader();
         // execute mapping file
         URL url = classLoader.getResource(mapPath);
@@ -47,10 +51,19 @@ public abstract class TestCore {
 
         File mappingFile = new File(mapPath);
         QuadStore rmlStore = QuadStoreFactory.read(mappingFile);
+
+        if (parentPath == null) {
+            parentPath = mappingFile.getParent();
+        }
+
         rmlStore.addQuads(extraQuads);
 
         return new Executor(rmlStore,
-                new RecordsFactory(mappingFile.getParent()), Utils.getBaseDirectiveTurtle(mappingFile));
+                new RecordsFactory(parentPath), Utils.getBaseDirectiveTurtle(mappingFile));
+    }
+
+    Executor createExecutor(String mapPath, String parentPath) throws Exception {
+        return createExecutor(mapPath, new ArrayList<>(), parentPath);
     }
 
     Executor createExecutor(String mapPath, FunctionLoader functionLoader) throws Exception {
@@ -90,6 +103,12 @@ public abstract class TestCore {
         }
     }
 
+    /**
+     * This method executes a mapping, compares it to the expected out, and returns the used Executor.
+     * @param mapPath The path of the mapping file.
+     * @param outPath The path of the file with the expected output.
+     * @return The Executor used to execute the mapping.
+     */
     public Executor doMapping(String mapPath, String outPath) {
         try {
             Executor executor = this.createExecutor(mapPath);
@@ -103,6 +122,31 @@ public abstract class TestCore {
         return null;
     }
 
+    /**
+     * This method executes a mapping, compares it to the expected out, and returns the used Executor.
+     * @param mapPath The path of the mapping file.
+     * @param outPath The path of the file with the expected output.
+     * @param parentPath The path of the folder where the Executor looks for files, such as CSV files.
+     * @return The Executor used to execute the mapping.
+     */
+    public Executor doMapping(String mapPath, String outPath, String parentPath) {
+        try {
+            Executor executor = this.createExecutor(mapPath, parentPath);
+            doMapping(executor, outPath);
+            return executor;
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            fail();
+        }
+
+        return null;
+    }
+
+    /**
+     * This method executes a mapping and compares with the expected output.
+     * @param executor The Executor that is used to execute the mapping.
+     * @param outPath The path of the file with the expected output.
+     */
     void doMapping(Executor executor, String outPath) throws Exception {
         QuadStore result = executor.execute(null);
         result.removeDuplicates();